wrt/requirements.txt
bidict==0.22.1
psycopg2-binary==2.9.6
python-engineio==4.4.1
python-socketio==5.8.0
websocket-client==1.5.1